corrects

+++ Project1/LogicApiTest/BallTest.cs
@@ -15,8 +15,8 @@ namespace LogicTest
         public void UpdatePositionTest()
         {
             Ball ball = new Ball();
-            float positionX = ball.x;
-            float positionY = ball.y;
+            double positionX = ball.x;
+            double positionY = ball.y;
             ball.updatePosition(530);
             Assert.AreEqual(ball.x, positionX + ball.xS);
             Assert.AreEqual(ball.y, positionY + ball.yS);
